+++ embulk-core/src/main/java/org/embulk/EmbulkEmbed.java
@@ -20,10 +20,18 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import java.util.function.Function;
+import org.embulk.config.CharsetJacksonModule;
+import org.embulk.config.ColumnJacksonModule;
 import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigLoader;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.DataSource;
+import org.embulk.config.LocalFileJacksonModule;
+import org.embulk.config.SchemaJacksonModule;
+import org.embulk.config.TimestampJacksonModule;
+import org.embulk.config.ToStringJacksonModule;
+import org.embulk.config.ToStringMapJacksonModule;
+import org.embulk.config.TypeJacksonModule;
 import org.embulk.exec.BulkLoader;
 import org.embulk.exec.ExecModule;
 import org.embulk.exec.ExecutionResult;
@@ -37,7 +45,6 @@ import org.embulk.exec.SystemConfigModule;
 import org.embulk.exec.TransactionStage;
 import org.embulk.jruby.JRubyScriptingModule;
 import org.embulk.spi.BufferAllocator;
-import org.embulk.spi.ColumnJacksonModule;
 import org.embulk.spi.DecoderPlugin;
 import org.embulk.spi.EncoderPlugin;
 import org.embulk.spi.ExecSessionInternal;
@@ -50,13 +57,6 @@ import org.embulk.spi.GuessPlugin;
 import org.embulk.spi.InputPlugin;
 import org.embulk.spi.OutputPlugin;
 import org.embulk.spi.ParserPlugin;
-import org.embulk.spi.SchemaJacksonModule;
-import org.embulk.spi.time.TimestampJacksonModule;
-import org.embulk.spi.type.TypeJacksonModule;
-import org.embulk.spi.unit.LocalFileJacksonModule;
-import org.embulk.spi.unit.ToStringJacksonModule;
-import org.embulk.spi.unit.ToStringMapJacksonModule;
-import org.embulk.spi.util.CharsetJacksonModule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
