+++ embulk-core/src/main/java/org/embulk/config/SchemaJacksonModule.java
@@ -1,4 +1,4 @@
-package org.embulk.spi;
+package org.embulk.config;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -15,6 +15,8 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
+import org.embulk.spi.Column;
+import org.embulk.spi.Schema;
 
 public final class SchemaJacksonModule extends SimpleModule {
     public SchemaJacksonModule() {
