+++ embulk-core/src/main/java/org/embulk/config/TypeJacksonModule.java
@@ -1,10 +1,11 @@
-package org.embulk.spi.type;
+package org.embulk.config;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import java.io.IOException;
+import org.embulk.spi.type.Type;
 
 public final class TypeJacksonModule extends SimpleModule {
     public TypeJacksonModule() {
