+++ embulk-core/src/test/java/org/embulk/EmbulkTestRuntime.java
@@ -9,9 +9,17 @@ import com.google.inject.Module;
 import java.util.Properties;
 import java.util.Random;
 import org.embulk.EmbulkEmbed;
+import org.embulk.config.CharsetJacksonModule;
+import org.embulk.config.ColumnJacksonModule;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.DataSourceImpl;
+import org.embulk.config.LocalFileJacksonModule;
 import org.embulk.config.ModelManager;
+import org.embulk.config.SchemaJacksonModule;
+import org.embulk.config.TimestampJacksonModule;
+import org.embulk.config.ToStringJacksonModule;
+import org.embulk.config.ToStringMapJacksonModule;
+import org.embulk.config.TypeJacksonModule;
 import org.embulk.exec.ExecModule;
 import org.embulk.exec.ExtensionServiceLoaderModule;
 import org.embulk.exec.SystemConfigModule;
@@ -19,19 +27,11 @@ import org.embulk.jruby.JRubyScriptingModule;
 import org.embulk.plugin.PluginClassLoaderFactory;
 import org.embulk.plugin.PluginClassLoaderFactoryImpl;
 import org.embulk.spi.BufferAllocator;
-import org.embulk.spi.ColumnJacksonModule;
 import org.embulk.spi.ExecAction;
 import org.embulk.spi.ExecInternal;
 import org.embulk.spi.ExecSessionInternal;
 import org.embulk.spi.MockFormatterPlugin;
 import org.embulk.spi.MockParserPlugin;
-import org.embulk.spi.SchemaJacksonModule;
-import org.embulk.spi.time.TimestampJacksonModule;
-import org.embulk.spi.type.TypeJacksonModule;
-import org.embulk.spi.unit.LocalFileJacksonModule;
-import org.embulk.spi.unit.ToStringJacksonModule;
-import org.embulk.spi.unit.ToStringMapJacksonModule;
-import org.embulk.spi.util.CharsetJacksonModule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
