+++ embulk-core/src/test/java/org/embulk/spi/type/TestTypeSerDe.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import java.io.IOException;
+import org.embulk.config.TypeJacksonModule;
 import org.junit.Test;
 
 public class TestTypeSerDe {
